move auth.ts into index.ts
